vo_vdpau: Add HEVC specific hacks to 'fix' the interleaving problem
This is a vicious hack, but if we double the height of the video
surface, the decoder, and the video mixer, then the decoder will
store the decoded frame in the space that the theoretical 'top field'
would occupy.

Then, if we tell the renderer to only render the top field, it will
end up rendering the correct frame that we wanted.

The disadvantage of this hack is that it prevents 4k video from
being decoded, as the decoder's maximum height is only just enough
for 4k.